Address comments from dustin
