+++ karma.conf.js
@@ -22,7 +22,6 @@ module.exports = function(config) {
       {pattern: 'src/main.js'},
       {pattern: 'src/**/*.js'},
       // ------------------------- Test Files ----------------------------------
-      'test/utils.js',
       'test/*_test.js',
     ],
     preprocessors: {'**/*.js': ['googmodule']},
